fix: conversion
Former-commit-id: 2e3d31223218dca599f1642854fdb710acf82854

+++ Core/Code/Extensions/EnumExtensions.cs
@@ -13,7 +13,7 @@ public static class EnumExtensions
     {
         try
         {
-            return BitOperations.PopCount(Convert.ToUInt64(flags));
+            return BitOperations.PopCount((ulong)Convert.ToInt64(flags));
         }
         catch (OverflowException ex)
         {
